Suppress studio supabase warning
This change suppress this message:

```
 ⚠ ./node_modules/.pnpm/@supabase+supabase-js@2.45.0_bufferutil@4.0.9_utf-8-validate@6.0.5/node_modules/@supabase/supabase-js/dist/module
Package @supabase/realtime-js can't be external
The request @supabase/realtime-js matches serverExternalPackages (or the default list).
The request could not be resolved by Node.js from the project directory.
Packages that should be external need to be installed in the project directory, so they can be resolved from the output files.
Try to install it into the project directory by running npm install @supabase/realtime-js from the project directory.

 ⚠ ./packages/supabase-driver/dist
Package @supabase/supabase-js can't be external
The request @supabase/supabase-js matches serverExternalPackages (or the default list).
The package resolves to a different version when requested from the project directory (2.50.0) compared to the package requested from the importing module (2.45.0).
Make sure to install the same version of the package in both locations.
```